changed log messages

+++ genweb/organs/browser/views.py
@@ -90,27 +90,28 @@ class Delete(BrowserView):
         action = self.request.form.get('action')
         itemid = self.request.form.get('item')
         portal_type = self.request.form.get('portal_type')
-        try:
-            if action == 'delete':
-                if '/' in itemid:
-                    # Es tracta de subpunt i inclou punt/subpunt a itemid (segon nivell)
-                    folder_path = '/'.join(self.context.getPhysicalPath()) + '/' + str('/'.join(itemid.split('/')[:-1]))
-                    itemid = str(''.join(itemid.split('/')[-1:]))
-                else:
-                    # L'objecte a esborrar es a primer nivell
-                    folder_path = '/'.join(self.context.getPhysicalPath())
+        if action == 'delete':
 
-                deleteItem = portal_catalog.searchResults(
-                    portal_type=portal_type,
-                    path={'query': folder_path, 'depth': 1},
-                    id=itemid)[0].getObject()
+            if '/' in itemid:
+                # Es tracta de subpunt i inclou punt/subpunt a itemid (segon nivell)
+                folder_path = '/'.join(self.context.getPhysicalPath()) + '/' + str('/'.join(itemid.split('/')[:-1]))
+                itemid = str(''.join(itemid.split('/')[-1:]))
+            else:
+                # L'objecte a esborrar es a primer nivell
+                folder_path = '/'.join(self.context.getPhysicalPath())
+
+            element = portal_catalog.searchResults(
+                portal_type=portal_type,
+                path={'query': folder_path, 'depth': 1},
+                id=itemid)
+
+            if element:
+                deleteItem = element[0].getObject()
                 with api.env.adopt_roles(['OG1-Secretari']):
                     api.content.delete(deleteItem)
                 portal_catalog = getToolByName(self, 'portal_catalog')
+                addEntryLog(self.context, None, _(u'Deleted via javascript'), deleteItem.Title() + ' - (' + self.request.form.get('item') + ')')
                 folder_path = '/'.join(self.context.getPhysicalPath())
-                addEntryLog(self.context, None, _(u'Deleted via javascript'), self.request.form.get('item'))
-
-                # agafo items ordenats
                 puntsOrdered = portal_catalog.searchResults(
                     portal_type=['genweb.organs.punt', 'genweb.organs.acord'],
                     sort_on='getObjPositionInParent',
@@ -125,23 +126,16 @@ class Delete(BrowserView):
                     if len(objecte.items()) > 0:
                         search_path = '/'.join(objecte.getPhysicalPath())
                         subpunts = portal_catalog.searchResults(
-                            portal_type=['genweb.organs.subpunt'],
+                            portal_type=['genweb.organs.subpunt', 'genweb.organs.acord'],
                             sort_on='getObjPositionInParent',
                             path={'query': search_path, 'depth': 1})
-
                         subvalue = 1
                         for value in subpunts:
                             newobjecte = value.getObject()
-                            newobjecte.proposalPoint = unicode(str(index) + str('.') + str(subvalue))
+                            newobjecte.proposalPoint = str(index) + str('.') + str(subvalue)
                             newobjecte.reindexObject()
                             subvalue = subvalue + 1
-
                     index = index + 1
-                # This line is only to bypass the CSRF WARNING
-                # WARNING plone.protect error parsing dom, failure to add csrf token to response for url ...
-                self.request.response.redirect(self.context.absolute_url())
-        except:
-            self.request.response.redirect(self.context.absolute_url())
 
 
 class Move(BrowserView):
